Splitting image utils tests
